<agrotis> adicionando data.sql, configurando hibernate, refatorando mapper

+++ src/main/java/com/agrotis/agrotis_backend/repository/PropriedadeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PropriedadeRepository extends JpaRepository<Propriedade, Integer> {
+public interface PropriedadeRepository extends JpaRepository<Propriedade, Long> {
 
 }
